% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationCalling.R
\name{baseCountsFromBamList}
\alias{baseCountsFromBamList}
\title{Create a mutationCalls object from a list of single-cell BAM files}
\usage{
baseCountsFromBamList(bamfiles, sites = "chrM:1-16659")
}
\arguments{
\item{bamfiles}{A character vector specifyign the bam file paths}

\item{sites}{Vector specifying genomic regions, defaults to the entire mitochondrial genome. Can be modified to query nuclear mutations}
}
\value{
A list of base count matrices which can serve as an input to \code{\link{mutationCallsFromBlacklist}} or \code{\link{mutationCallsFromCohort}}
}
\description{
Uses the \code{deepSNV} package to count nucleotide frequencies at every position in the mitochondrial genome for every cell and passes the result to the \code{\link{mutationCallsFromDeepSNV}} function for filtering variants.
}
