% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationCalling.R
\name{baseCountsFromSingleBam}
\alias{baseCountsFromSingleBam}
\title{Create a mutationCalls object from a list of single-cell BAM files}
\usage{
baseCountsFromSingleBam(bam, sites = "chrM:1-16659", tag = "XQ:C")
}
\arguments{
\item{bam}{Path to the bam file}

\item{sites}{Vector specifying genomic regions, defaults to the entire mitochondrial genome. Can be modified to query nuclear mutations}

\item{tag}{Name of the bam file tag}
}
\value{
A list of base count matrices which can serve as an input to \code{\link{mutationCallsFromBlacklist}} or \code{\link{mutationCallsFromCohort}}
}
\description{
More explanations of what happens
}
