% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blacklists}
\alias{blacklists}
\title{Mitochondrial blacklist}
\format{A list with three entries:
#' \itemize{
  \item \emph{three} **explanations ben**
  \item \emph{mutaseq} **explanations ben**
  \item \emph{masked} **explanations ben**
}}
\usage{
blacklists
}
\description{
Blacklist of variants that are likely not true somatic mutations
}
\keyword{datasets}
