% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clusterMetaclones}
\alias{clusterMetaclones}
\title{Cluster mutations into clones}
\usage{
clusterMetaclones(mutcalls, nclust = 3)
}
\arguments{
\item{mutcalls}{mutcalls object of class \code{\link{mutationCalls}} for which \code{\link{muta_cluster}} has been run}

\item{nclust}{number of metaclones expected}
}
\description{
PhISCS orders all mutations into a hierarchical mutational tree; in many cases, the exact order of the acquisition of individual mutations in not unanimously determined from the data. This function computes the change in likelihood of the infered clonal assignment if two mutations are merged into a clone. Hierarchical clustering is then used to determine the clonal structure. The result is visualized and should be fine-tuned using the \code{nclust} parameter.
}
