% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\alias{M_P1}
\alias{N_P1}
\alias{M_P2}
\alias{N_P2}
\title{Allele Count tables}
\format{
a data frame of variable sites (columns) across single cells (rows)

An object of class \code{data.frame} with 1430 rows and 16 columns.

An object of class \code{data.frame} with 1430 rows and 16 columns.

An object of class \code{data.frame} with 1066 rows and 22 columns.

An object of class \code{data.frame} with 1066 rows and 22 columns.
}
\usage{
M_P1

N_P1

M_P2

N_P2
}
\description{
M: Mutant allele counts; N: Reference allele counts. P1: Patient 1; P2: Patient 2
}
\keyword{datasets}
