% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationCalling.R
\name{mutationCallsFromBamList}
\alias{mutationCallsFromBamList}
\title{Create a mutationCalls object from a list of single-cell BAM files}
\usage{
mutationCallsFromBamList(bamfiles)
}
\arguments{
\item{bamfiles}{A character vector specifyign the bam file paths}
}
\value{
An object of class \code{\link{mutationCalls}}
}
\description{
More explanations of what happens
}
