% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationCalling.R
\name{mutationCallsFromBlacklist}
\alias{mutationCallsFromBlacklist}
\title{Create a mutationCalls object from nucleotide base calls using a blacklist (single individual)}
\usage{
mutationCallsFromBlacklist(BaseCounts, lim.cov = 20, min.af = 0.2,
  min.num.samples = 0.01 * length(BaseCounts),
  min.af.universal = min.af, universal.var.cells = 0.95 *
  length(BaseCounts), blacklists.use = blacklists, max.var.na = 0.5,
  max.cell.na = 0.95, ...)
}
\arguments{
\item{BaseCounts}{A list of base call matrices (one matrix per cell) as produced by \code{\link{baseCountsFromBamList}}}

\item{lim.cov}{Minimal coverage required per cell for a cell to be classified as covered}

\item{min.af}{Minimal allele frequency for a cell to be classified as mutant}

\item{min.num.samples}{Minimal number of cells required to be classified as covered and mutant according to the thresholds set in \code{lim.cov} and \code{min.af}. Usually specified as a fraction of the total number of cells.}

\item{min.af.universal}{Minimal allele frequency for a cell to be classified as mutant, in the context of removing universal variants. Defaults to \code{min.af}, but can be set to lower values.}

\item{universal.var.cells}{Maximum number of cells required to be classified as mutant according to the threshold set in \code{min.af.universal}.  Usually specified as a fraction of the total number of cells; serves to avoid e.g. germline variants.}

\item{max.var.na}{Final filtering step: Remove all mutations with no coverage in more than this fraction of cells}

\item{max.cell.na}{Final filtering step: Remove all cells with no coverage in more than this fraction of mutations}

\item{...}{Parameters passed to \code{\link{mutationCallsFromMatrix}}}

\item{blacklists}{Blacklists to use **explanations ben**}
}
\value{
An object of class \code{\link{mutationCalls}}
}
\description{
Identifies relevant mitochondrial somatic variants from raw counts of nucleotide frequencies. Applies two sets of filters: In the first step, filters on coverage and minimum allele frequency to exclude potentially noisy variants; in the second step, filters against a blacklist of variants that were observed in several individuals and that therefore are unlikely to represent true somatic variants (e.g. RNA editing events). These blacklists are created using \code{\link{mutationCallsFromCohort}}
}
