% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationCalling.R
\name{mutationCallsFromCohort}
\alias{mutationCallsFromCohort}
\title{Create a mutationCalls objects from nucleotide base calls and defines a blacklist (cohort)}
\usage{
mutationCallsFromCohort(
  BaseCounts,
  patient,
  MINREADS = 5,
  MINCELL = 20,
  MINFRAC = 0.1,
  MINCELLS.PATIENT = 10,
  MINRELATIVE.PATIENT = 0.01,
  MINRELATIVE.OTHER = 0.1
)
}
\arguments{
\item{BaseCounts}{A list of base call matrices (one matrix per cell) as produced by \code{\link{baseCountsFromSingleBam}} or \code{\link{baseCountsFromBamList}}}

\item{patient}{A character vector associating each cell / entry in the \code{BaseCount} list with a patient}

\item{MINREADS}{Minimum number of reads on a site in a single cell to qualify the site as covered}

\item{MINCELL}{Minimum number of cells across the whole data set to cover a site}

\item{MINFRAC}{Fraction of reads on the mutant allele to provisionally classify a cell as mutant}

\item{MINCELLS.PATIENT}{Minimum number of mutant cells per patient to classify the mutation as relevant in that patient, AND}

\item{MINRELATIVE.PATIENT}{Minimum fraction of mutant cells per patient to classify the mutation as relevant in that patient}

\item{MINRELATIVE.OTHER}{Minimum fraction of mutant cells identified in a second patient for the mutation to be excluded}
}
\value{
A list of \code{\link{mutationCalls}} objects (one for each \code{patient}) and an entry \code{blacklist} containing a blacklist of sites with variants in several individuals
}
\description{
Identifies relevant mitochondrial somatic variants from raw counts of nucleotide frequencies measured in single cells from several individuals. Applies two sets of filters: In the first step, filters on coverage to include potentially noisy variants; in the second step, compares allele frequencies between patients to remove variants that were observed in several individuals and that therefore are unlikely to represent true somatic variants (e.g. RNA editing events). The blacklist derived from the MutaSeq dataaset is available in \code{\link{blacklist}} and can be used on single individuals using \code{\link{mutationCallsFromBlacklist}}
}
