% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{mutationCallsFromMatrix}
\alias{mutationCallsFromMatrix}
\title{mutationCalls constructor}
\usage{
mutationCallsFromMatrix(
  M,
  N,
  cluster = NULL,
  metadata = data.frame(row.names = rownames(M)),
  binarize = function(M, N) {     alleleRatio <- M/(M + N)     apply(alleleRatio, 2,
    function(x) ifelse(is.na(x), "?", ifelse(x > 0.05, "1", "0"))) }
)
}
\arguments{
\item{M}{A matrix of read counts mapping to the \emph{mutant} allele. Columns are genomic sites and rows and single cells.}

\item{N}{A matrix of read counts mapping to the \emph{mutant} allele. Columns are genomic sites and rows and single cells.}

\item{cluster}{If \code{NULL}, only mutations with coverage in 20% of the cells or more will be used for the clustering, and all other mutations will be used for cluster annotation only. Alternatively, a boolean vector of length \code{ncol(M)} that specifies the desired behavior for each genomic site.}

\item{binarize}{A function that turns M and N into a ternary matrix of mutation calls in single cells, where where -1 signfiies reference, 0 signifies dropout, and 1 signifies wild-type. The default was found to work well on mitochondrial data.}
}
\value{
An object of class \code{\link{mutationCalls}}.
}
\description{
To be used when allele-specific count matrices are available.
}
