% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationCalling.R
\name{mutationCallsFromSingleBam}
\alias{mutationCallsFromSingleBam}
\title{Create a mutationCalls object from a list of single-cell BAM files}
\usage{
mutationCallsFromSingleBam(bam, tag = "XQ:C")
}
\arguments{
\item{bam}{Path to the bam file}

\item{tag}{Name of the bam file tag}
}
\value{
An object of class \code{\link{mutationCalls}}
}
\description{
More explanations of what happens
}
