% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{plotTree}
\alias{plotTree}
\title{Plot clonal tree}
\usage{
plotTree(mutcalls, file = "mytree.ps")
}
\arguments{
\item{mutcalls}{object of class \code{\link{mutationCalls}}.}

\item{file}{file name of postscript output file}
}
\description{
Uses graphviz to plot the clonal tree
}
