% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{quick_cluster}
\alias{quick_cluster}
\title{Quick clustering of mutations}
\usage{
quick_cluster(mutcalls, binarize = F, drop_empty = T, ...)
}
\arguments{
\item{mutcalls}{object of class \code{\link{mutationCalls}}.}

\item{binarize}{If \code{FALSE}, will use raw allele frequencies for the clustering. If \code{TRUE}, will use binarized mutation/reference/dropout calls. Alternatively, provide a function top create binarized mutation/reference/dropout calls from the matrices M and N, see \code{\link{mutationCallsFromMatrix}}}

\item{...}{Parameters passed to \code{\link{pheatmap::pheatmap}}}
}
\value{
The result of running \code{\link{pheatmap::pheatmap}}
}
\description{
Performs a quick hierarchical clustering on a object of class \code{\link{mutationCalls}}. See \code{\link{muta_cluster}} for an alternative that infers mutational trees and uses sound models of dropout.
}
