% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationCalling.R
\name{baseCountsFromBamList}
\alias{baseCountsFromBamList}
\title{Create a list object from a list of single-cell BAM files where each contains a matrix of the of AGCT nt counts at chosen sites}
\usage{
baseCountsFromBamList(bamfiles, sites = "chrM:1-16659", ncores = 20)
}
\arguments{
\item{bamfiles}{A character vector specifying the bam file paths}

\item{sites}{Vector specifying genomic regions, defaults to the entire mitochondrial genome.}

\item{ncores}{Number of threads to use for the computation}
}
\value{
A list of base count matrices which can serve as an input to \code{\link{mutationCallsFromBlacklist}} or \code{\link{mutationCallsFromCohort}}
}
\description{
Uses the \code{deepSNV} package to count nucleotide frequencies at every position in the mitochondrial genome for every cell and passes the result to the \code{\link{mutationCallsFromDeepSNV}} function for filtering variants.
}
