% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{plotClones}
\alias{plotClones}
\title{Plot clonal assignment of single cells}
\usage{
plotClones(mutcalls, what = "alleleRatio", ...)
}
\arguments{
\item{mutcalls}{object of class \code{\link{mutationCalls}}.}

\item{...}{any arguments passed to \code{\link{pheatmap::pheatmap}}}
}
\description{
Creates a heatmap of single cell mutation calls, clustered using PhISCS.
}
